Conflicts:
	iOS/NetplayController.m
	iOS/NetplayGameKit.h
	make-ios-sim.sh
	xcode/MAME4iOS/MAME4iOS copy-Info.plist
	xcode/MAME4iOS/MAME4iOS.xcodeproj/project.pbxproj

+++ iOS/NetplayController.m
@@ -139,6 +139,7 @@ static void netplay_warn_callback(char *msg)
                                                               target: emuController  action:  @selector(done:) ];
     self.navigationItem.rightBarButtonItem = button;
     [button release];
+    
 }
 
 - (NSInteger)numberOfSectionsInTableView:(UITableView *)tableView {
@@ -173,7 +174,7 @@ static void netplay_warn_callback(char *msg)
 
 - (UITableViewCell *)tableView:(UITableView *)tableView cellForRowAtIndexPath:(NSIndexPath *)indexPath {
     
-    NSString *cellIdentifier = [NSString stringWithFormat: @"%d:%d", (int)[indexPath indexAtPosition:0], (int)[indexPath indexAtPosition:1]];
+    NSString *cellIdentifier = [NSString stringWithFormat: @"%lu:%lu", (unsigned long)[indexPath indexAtPosition:0], (unsigned long)[indexPath indexAtPosition:1]];
     UITableViewCell *cell = [tableView dequeueReusableCellWithIdentifier:cellIdentifier];
     
     netplay_t *handle = netplay_get_handle();
@@ -353,6 +354,7 @@ static void netplay_warn_callback(char *msg)
                 return;
             
             [self startSocket];
+//            [self startGamekit];
         }
         else
         {            
@@ -372,11 +374,12 @@ static void netplay_warn_callback(char *msg)
         {
             if(![self ensureWIFI])
                 return;
-            
+
             if(![self ensurePeerAddr])
                 return;
             
             [self joinSocket];
+//            [self joinGamekit];
         }
         else
         {
@@ -545,7 +548,6 @@ static void netplay_warn_callback(char *msg)
 }
 
 -(bool)ensureBluetooth{
-    
     bool first = false;
     if(btMgr==nil)
     {
